+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Highlight delete and remove actions in red
 - added dedicated routes for statistics and bar tabs
 - Category chips in statistics are now only shown when corresponding data is available
+- refactor statistics screen files
 
 ### Fixed
 - Fix warning in web regarding missing emoji font subset